pyproject.toml
# Copyright 2021 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[project]
name = "marketing-analytics-jumpstart"
version = "1.0.0"
description = "Marketing Analytics Jumpstart"
authors = [{name = "Marketing Analytics Solutions Architects", email = "ma-se@google.com"}]
license = "Apache 2.0"
readme = "README.md"
requires-python = ">=3.9,<3.12.0"

[tool.poetry]
name = "marketing-analytics-jumpstart"
version = "1.0.0"
description = "Marketing Analytics Jumpstart"
authors = ["Marketing Analytics Solutions Architects <ma-se@google.com>"]
license = "Apache 2.0"
readme = "README.md"
packages = [{include = "python"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.12.0"
#google-cloud-aiplatform = "1.52.0"
google-cloud-aiplatform = "1.77.0"
shapely = "<2.0.0"
google-cloud = "^0.34.0"
jinja2 = ">=3.0.1,<4.0.0"
pip = "23.3.2"
invoke = "2.2.0"
## pyinvoke = "1.0.4"
pre-commit = ">=2.14.1,<3.0.0"
pandas = "1.5.3"
google-cloud-bigquery = "3.21.0"
google-cloud-bigquery-connection = "1.17.0"
#google-cloud-pipeline-components = "1.0.33"
google-cloud-pipeline-components = "2.6.0"
google-auth = "^2.14.1"
google-cloud-storage = "^2.6.0"
kfp = "2.4.0"
## Fixing this error: https://stackoverflow.com/questions/76175487/sudden-importerror-cannot-import-name-appengine-from-requests-packages-urlli
#kfp = "2.0.0-rc.2"
#kfp = {version = "2.0.0-b12", allow-prereleases = true}
#kfp = {version = "2.0.0-b16", allow-prereleases = true}
kfp-server-api = "2.0.5"
#kfp-server-api = "2.0.0-rc.1"
#kfp-server-api = "2.0.0.a6"
#kfp-server-api = "2.0.0b1"
urllib3 = "1.26.20"
toml = "0.10.2"
docker = "^6.0.1"
db-dtypes = "1.3.1"
optuna = "3.6.1"
scikit-learn = "1.5.0"
#plotly = "5.16.0" 
#matplotlib= "3.7.2" 
#seaborn = "0.12.2"
ma-components = {path = "python/base_component_image/", develop = true}
google-cloud-pubsub = "2.27.2"
#google-analytics-admin = "0.17.0"
google-analytics-admin = "0.23.3"
google-analytics-data = "^0.18.0"
pyarrow = "15.0.2"
google-auth-oauthlib = "^1.2.1"
oauth2client = "^4.1.3"
google-cloud-core = "^2.4.1"
sympy="1.13.3"
google-cloud-resource-manager="1.14.0"
google-cloud-service-usage = ">=1.13.0"
google-cloud-iam = ">=2.18.0"

[tool.poetry.group.component_vertex.dependencies]
google-cloud-aiplatform = "1.77.0"
shapely = "<2.0.0"
toml = "0.10.2"

[tool.poetry.scripts]
ga4-setup = "python.ga4_setup.setup:entry"
bq-set-permissions = "python.data_store.setup:set_bq_permission"
service-apis-setup = "python.deployment.enable_service_apis:service_apis_setup"
deployer-roles-check = "python.deployment.deployer_iam_role_manager:check_missing_roles"

[tool.poetry.group.test.dependencies]
pytest = "7.4.4"
pytest-env = "0.8.2"
pytest-mock = "3.14.0"
pytest-variables = {extras = ["yaml"], version = "^2.0.0"}
coverage = {extras = ["toml"], version = "^6.5.0"}
pytest-cov = "^4.0.0"
pytest-xdist = "^3.0.2"

[tool.poetry.group.dev.dependencies]
pip = "23.3.2"
invoke = "2.2.0"
pre-commit = ">=2.14.1,<3.0.0"
black = "22.12.0"
flake8 = "5.0.4"
flake8-annotations = "2.9.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest-mock>=3.14.0",
    "pytest>=8.3.4",
    "google-cloud-workflows>=1.16.0",
    "google-cloud-dataform>=0.5.15",
]

[tool.pytest.ini_options]
testpaths = ["python/pipelines"]
markers = [
    "unit",
    "inte",
    "pipeline_run",
    "feature_eng_pl",
    "training_pl",
    "customer_seg_pl",
    "current",
    "curl",
    "schedule",
    "dataform"
]
log_cli=true
log_level="INFO"

[tool.coverage.run]
include = ["*/python/*"]
omit = ["*/tests/*"]
parallel = true

[tool.coverage.report]
fail_under = 70
show_missing = true
skip_empty= true

[tool.uv.workspace]
members = ["python/lookerstudio"]
